refactor(profileformgr): refactored groomer form files to use context
