+++ src/state/contexts/CustomersContext.js
@@ -3,8 +3,20 @@ import React, { createContext, useState } from 'react';
 export const CustomersContext = createContext({});
 
 const CustomersProvider = ({ children }) => {
+  const [custInfo, setCustInfo] = useState({});
+  const [updated, setUpdated] = useState(false);
+
   return (
-    <CustomersContext.Provider value={{}}>{children}</CustomersContext.Provider>
+    <CustomersContext.Provider
+      value={{
+        custInfo,
+        setCustInfo,
+        updated,
+        setUpdated,
+      }}
+    >
+      {children}
+    </CustomersContext.Provider>
   );
 };
