+++ src/state/contexts/RootContext.js
@@ -3,15 +3,18 @@ import CustomersProvider from './CustomersContext';
 import GroomersProvider from './GroomersContext';
 import UsersContext from './UsersContext';
 import APIContext from './APIContext';
+import FormContext from './FormContext';
 
 export const RootProvider = ({ children }) => {
   return (
-    <UsersContext>
-      <GroomersProvider>
-        <CustomersProvider>
-          <APIContext>{children}</APIContext>
-        </CustomersProvider>
-      </GroomersProvider>
-    </UsersContext>
+    <FormContext>
+      <UsersContext>
+        <GroomersProvider>
+          <CustomersProvider>
+            <APIContext>{children}</APIContext>
+          </CustomersProvider>
+        </GroomersProvider>
+      </UsersContext>
+    </FormContext>
   );
 };
